requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
tweepy==4.5.0
urllib3==1.26.8
Mastodon.py==1.5.2